--- components/Question.js
@@ -1,13 +1,0 @@
-import { Question } from "./styled/Question.styled";
-
-const QuestionSection = () => {
-  return (
-    <>
-      <Question>
-          
-      </Question>
-    </>
-  );
-};
-
-export default QuestionSection;
