+++ components/styled/Card.styled.js
@@ -1,0 +1,21 @@
+import styled from "styled-components";
+
+export const Question = styled.div`
+  width: 464px;
+  height: 515px;
+  border-radius: 24px;
+  background-color: #fff;
+  & > div > #question-title {
+    width: 273px;
+    margin-top: -50px;
+    font-family: "Poppins", sans-serif;
+    font-style: normal;
+    font-weight: bold;
+    font-size: 36px;
+    line-height: 54px;
+    text-transform: uppercase;
+    color: #f2f2f2;
+  }
+  & > div > #question-world-image {
+  }
+`;
