+++ components/styled/Photo.styled.js
@@ -1,0 +1,9 @@
+import Image from "next/image";
+
+import styled from "styled-components";
+
+export const Photo = styled(Image)`
+  position: relative;
+  top: -75px;
+  left: 30px;
+`;
