--- components/styled/Question.styled.js
@@ -1,8 +1,0 @@
-import styled from "styled-components";
-
-export const Question = styled.div`
-  width: 464px;
-  height: 515px;
-  border-radius: 24px;
-  background-color: #fff;
-`;
